Some fix according to andrewr224 comments
